Separate Bipartite class out
